#162 Move `PageSwitcher` to a separate file
