+++ iolanta/facets/textual_browser/location.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Location:
+    """Unique ID and IRI associated with it."""
+
+    page_id: str
+    url: str
